Unload plugin from memory before removing file Update contact plugin for error text handling

+++ linphone-app/src/components/contacts/ContactsImporterPluginsManager.cpp
@@ -67,8 +67,12 @@ void ContactsImporterPluginsManager::openNewPlugin(){
 		QString path = Utils::coreStringToAppString(Paths::getPluginsContactsDirPath())+fileInfo.fileName();
 		if( QFile::exists(path)){
 			doCopy = QMessageBox::question(nullptr, "Importing Address Book Connector", "The plugin already exists. Do you want to overwrite it?", QMessageBox::Yes, QMessageBox::No);
-			if( doCopy == QMessageBox::Yes)
-				QFile::remove(path);
+			if( doCopy == QMessageBox::Yes){
+				if(!QFile::remove(path)){
+					if(QPluginLoader(path).unload())// Try to unload plugin
+						QFile::remove(path);
+				}
+			}
 		}
 		if(doCopy != QMessageBox::Yes || !QFile::copy(fileName, path))
 			qWarning() << "Cannot copy plugin";
